requirements.txt
psycopg2-binary
celery
redis==3.3.11
boto3==1.10.45
django-storages==1.8